spike: refactor reload workspace cmd
