+++ packages/plugin-core/src/workspace/nativeWorkspace.ts
@@ -1,5 +1,6 @@
 import {
   DendronConfig,
+  DVault,
   DWorkspaceV2,
   WorkspaceType,
 } from "@dendronhq/common-all";
@@ -9,9 +10,11 @@ export class DendronNativeWorkspace implements DWorkspaceV2 {
   public wsRoot: string;
   public type = WorkspaceType.NATIVE;
   public config: DendronConfig;
+  public vaults: DVault[];
 
   constructor({ wsRoot }: { wsRoot: string }) {
     this.wsRoot = wsRoot;
     this.config = DConfig.getOrCreate(wsRoot);
+    this.vaults = this.config.vaults;
   }
 }
